part4 - possibly the most jank tests ever that 100% need to be refactored

+++ blog/controllers/blogs.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 const router = express.Router();
 
 router.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
@@ -15,15 +16,17 @@ router.get("/:id", async (request, response) => {
 
 router.post("/", async (request, response) => {
   let { title, author, url, likes } = request.body;
-  // if (!title || !url) {
-  //   return response
-  //     .status(400)
-  //     .json({ error: "Title and author fields must be defined" });
-  // }
-  // if (!likes) likes = 0;
-  const blog = new Blog({ title, author, url, likes });
-  const savedNote = await blog.save();
-  response.status(201).json(savedNote);
+
+  const user = request.user;
+  if (!user)
+    return response.status(401).json({ error: "Please provide a valid token" });
+
+  const blog = new Blog({ title, author, url, likes, user: user.id });
+  const savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+  response.status(201).json(savedBlog);
 });
 
 router.put("/:id", async (request, response) => {
@@ -35,8 +38,18 @@ router.put("/:id", async (request, response) => {
 });
 
 router.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blogToRemove = await Blog.findById(request.params.id);
+
+  const user = request.user;
+
+  if (user.id.toString() === blogToRemove.user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    return response.status(204).end();
+  }
+
+  return response
+    .status(401)
+    .json({ error: "Not authorized to delete this blog post" });
 });
 
 module.exports = router;
